+++ CHANGELOG.md
@@ -3,6 +3,17 @@ All notable changes to this project will be documented in this file. This change
 
 ## [Unreleased]
 
+## 0.9.0 - 2022-03-25
+
+### Added
+- Documentation for quick-start, install, and typical API usage
+
+### Changed
+- Migrated from SymSpell to Lucene using Jena's inbuilt fulltext search
+- Edit distance no longer configurable
+
+### Removed
+- SymSpell dependency
 ## 0.1.0 - 2020-08-08
 ### Added
 - First rough version.
@@ -10,4 +21,5 @@ All notable changes to this project will be documented in this file. This change
 - SymSpell implementation by [@MighTguY ](https://github.com/MighTguY/customized-symspell/)
 
 [Unreleased]: https://github.com/AberystwythSystemsBiology/sog/compare/v0.1.0...HEAD
+[0.9.0]: https://github.com/AberystwythSystemsBiology/sog/releases/tag/v0.9.0
 [0.1.0]: https://github.com/AberystwythSystemsBiology/sog/releases/tag/v0.1.0